model-service/requirements.txt
flask
prometheus_client
kafka-python
scikit-learn
numpy
pandas
joblib
requests
flask-cors
